Fix dev tool
I encountered the following error. This is my attempt to fix it.:
 + python3 -m venv venv
The virtual environment was not created successfully because ensurepip is not
available.  On Debian/Ubuntu systems, you need to install the python3-venv
package using the following command.
    apt install python3.12-venv
You may need to use sudo with that command.  After installing the python3-venv
package, recreate your virtual environment.
[1729565820]
[1729565820] Failing command: /workspace/comment-sense/venv/bin/python3
[1729565820]